Handle configuration of the button and update README
